bot_api/requirements.txt
anyio==4.3.0
APScheduler==3.10.4
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
psycopg2-binary
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
python-telegram-bot-calendar==1.0.5
pytz==2024.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
